Create perfecr number
Enter a Number : 
6
The Entered Number is a PERFECT Number. 

Enter a Number : 
17
The Entered Number is not a PERFECT Number. 

Enter a Number : 
26!
Enter a Valid Number

Enter a Number : 
143
The Entered Number is not a PERFECT Number. 

Enter a Number : 
84.1
Enter a Valid Number

Enter a Number : 
-963
The Entered Number is not a PERFECT Number.